requirements.txt
numpy==1.19.2
pandas==1.2.2
plotly==4.14.3
matplotlib==3.3.4
shap==0.38.1
xgboost==1.3.3
umap_learn==0.5.1
streamlit==0.77.0
Pillow==8.1.1
scikit_learn==0.24.1
umap==0.1.1